stable/linux-dmabuf: allow compositors to advertise multiple devices
With multi gpu systems, if the client is rendering on a device other than the
main_device advertised by the compositor, there's no way for the client to know
whether or not the import will be successful except for allocating a buffer and
attempting to import the buffer into the target device, and if that fails,
trying again with different formats, modifiers and potentially allocation flags.
Similarly, compositors that want to supoprt buffers from non-main devices need
to attempt importing the buffer to every device they support. In some cases,
this import may even trigger expensive operations such as the buffer being
copied into system memory by the kernel.

This protocol addition allows compositors to explicitly advertise support for
multiple devices on the system, and specify which formats and modifiers they
can sample from.
For buffer imports, the client explicitly specifies a device for the compositor
to import the buffer to, which allows the compositor to avoid additional copies.

Signed-off-by: Xaver Hugl <xaver.hugl@kde.org>