color: remove bt1361 TF
This historical recommendation has been withdrawn by ITU in 2015.

The "conventional color gamut" portion is equivalent to BT.709 which is
to be decoded with BT.1886.

The "extended color gamut" portion that uses negative RGB values is
unclear how it should be decoded. I haven't seen any recommendation to
that yet. BT.1886 clamps negative values to zero at zero black level
lift.

Therefore I believe it is unlikely to need this in the protocol.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>