color: rename TF bt709 to bt1886
BT.709 defines a camera characteristic (OETF). A camera characteristic
is irrelevant for video distribution and broadcast, because the signal
is color graded while watching the result on a reference display
(producing a display-referred signal). BT.709 does not define any
display characteristic, but refers to BT.1886 in a footnote.

BT.601 and BT.2020 are similar to BT.709 in this respect.

The meaning of this enum entry remains the same, only its name is
changed to refer to an actual display characteristic definition.

H.273 mentions more references. Two of them are historical and withdrawn
by the ITU in 2015. BT.1700 NTSC refers to SMPTE ST 170 which is unclear
about the display characteristic: it claims the display EOTF is the
exact inverse of the given OETF. This does not match BT.1886 even though
H.273 implies it does. Hence these are not listed in our enum entry.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>