color: elaborate on luminance
'set_luminances' request is intended to reflect the viewing conditions,
not only the display emission levels. Mention explicitly that it
includes display light leak and ambient flare, as these cannot be
avoided while viewing.

Ambient flare is physically additive. I was thinking of LCD light
leaks as the minimum display emission, which I presume to be roughly
additive. Effects that are additive in non-optical spaces, e.g.
black-level lift in electrical space, would need their own
parametrization to account for. To keep things simpler, we assume
non-optical additives are zero.

The same definition is applied to the MDCV and target color volume. I
believe it would be confusing otherwise. I can only hope that this is
compatible with SMPTE ST 2086.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>