color: define global_remove behavior
https://gitlab.freedesktop.org/wayland/wayland-protocols/-/merge_requests/14#note_2756612
https://gitlab.freedesktop.org/wayland/wayland/-/issues/484

Toolkit authors are reluctant to handle the global removal, because on
most compositors the removal will never happen. OTOH, there is a
compositor author who wants to be able to change color-management
capabilities at runtime due to dynamically switching renderers.

This new wording strongly suggests the global is never removed to let
toolkits avoid all the code that removal tracking would need. Removal is
still technically allowed, but it is not generally expected that clients
would handle it.

Signed-off-by: Pekka Paalanen <pekka.paalanen@collabora.com>